requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
jinja2
numpy
soundfile
pythainlp
f5-tts-th
scipy
librosa